Pulled template's map generation code out into its own class, and commented the rest of the code more extensively to detail common customization points.
